Time: 18 ms (42.05%), Space: 8.5 MB (41.34%) - LeetHub

+++ 451-sort-characters-by-frequency/451-sort-characters-by-frequency.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string frequencySort(string s) {
-        unordered_map<char,long> freq;
+        unordered_map<char,int> freq;
         priority_queue<pair<int,char>> maxh;
         string ans;
